+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpecBase.kt
@@ -36,7 +36,7 @@ abstract class IntelliJPluginSpecBase {
         ?: throw GradleException("'test.markdownPlugin.version' isn't provided")
     val dir: File by lazy { createTempDirectory("tmp").toFile() }
 
-    private val gradleProperties = file("gradle.properties")
+    val gradleProperties = file("gradle.properties")
     val buildFile = file("build.gradle")
     val pluginXml = file("src/main/resources/META-INF/plugin.xml")
     val buildDirectory = File(dir, "build")
@@ -52,8 +52,8 @@ abstract class IntelliJPluginSpecBase {
                 id 'org.jetbrains.intellij'
                 id 'org.jetbrains.kotlin.jvm' version '$kotlinPluginVersion'
             }
-            sourceCompatibility = 1.8
-            targetCompatibility = 1.8
+            sourceCompatibility = 11
+            targetCompatibility = 11
             repositories {
                 mavenCentral()
             }